backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.4.2
mysql-connector-python==9.3.0
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
